Move github_handle logic into model
Mixin was only used once
